pr to talk about the merge
This makes it so even if you have 100% Unlocked you still have to hotwire / can not drive the vehicle. Some players do not want this. 
Therefore I have made a way that I have had working to disable it, and I assumed it should be a configurable option. 

Let's Discuss?